+++ templates/ffcli/grpc/proto/swissknife.proto
