+++ templates/ffcli/grpc/templates/cmd/package/main.go.tmpl
@@ -46,21 +46,29 @@ import (
     "flag"
     "fmt"
     "log"
+	"net"
     "os"
 
-
+	"google.golang.org/grpc"
+    "$[ADAPTERKIT_GOMOD]/pkg/{{$package}}_gen"
+    
+    {{$package}} "$[ADAPTERKIT_LOGIC_PACKAGE]"
     "github.com/peterbourgon/ff/v3"
     "github.com/peterbourgon/ff/v3/ffcli"
 )
 
+func initSvc() {{$package}}.{{$package | title}}SvcServer {
+        return {{$package}}.New()
+}
+
 func main() {
-    if err := {{$package}}(os.Args[1:]); err != nil {
+    if err := {{$package}}Run(os.Args[1:]); err != nil {
         log.Fatal(err)
     }
     return
 }
 
-func {{$package}}(args []string) error {
+func {{$package}}Run(args []string) error {
     rootFlagSet := flag.NewFlagSet("{{$package}}", flag.ExitOnError)
 
     root := &ffcli.Command{
@@ -69,8 +77,9 @@ func {{$package}}(args []string) error {
         Options:    []ff.Option{ff.WithEnvVarNoPrefix()},
         Subcommands: []*ffcli.Command{
 {{- range $m := .Service.Method}}
-            {{$m.Name}}(),
+            {{$m.Name | lowerFirst}}(),
 {{- end}}
+            server(),
         },
         Exec: func(_ context.Context, _ []string) error {
             return flag.ErrHelp
@@ -84,7 +93,7 @@ func {{$package}}(args []string) error {
 {{- range $m := .Service.Method}}
 {{- $p := $m.Name | lowerFirst}}
 {{- $t := $m.InputType | getMessageType $.File}} 
-func {{$m.Name}}() *ffcli.Command {
+func {{$p}}() *ffcli.Command {
     {{template "iterateOnFieldsName&Type1" $t}}
     {{$p}}FlagSet := flag.NewFlagSet("{{$m.Name | lowerFirst}}", flag.ExitOnError)
     {{template "iterateOnFieldsFfcliName&Type" $x := list $t $p}}
@@ -104,4 +113,24 @@ func {{$m.Name}}() *ffcli.Command {
         },
     }
 }
+
 {{end}}
+func server() *ffcli.Command {
+	return &ffcli.Command{
+		Name:       "server",
+		ShortUsage: "{{$package}} start",
+		Options:    []ff.Option{ff.WithEnvVarNoPrefix()},
+		Exec: func(_ context.Context, _ []string) error {
+			lis, err := net.Listen("tcp", {{/* MUST BE REPLACED BY ARG GEST */}}"127.0.0.1:9314")
+			if err != nil {
+				return err
+			}
+			grpcServer := grpc.NewServer()
+
+			{{$package}}.Register{{$package | title}}SvcServer(grpcServer, initSvc())
+
+			fmt.Println("starting demo-mod_gen server on port 9314:")
+			return grpcServer.Serve(lis)
+		},
+	}
+}
